requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Task Queue
celery==5.3.4
kombu==5.3.4

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# WebSocket
websockets==12.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
ta==0.10.2

# Notifications
aiosmtplib==3.0.1
python-telegram-bot==20.7

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
factory-boy==3.3.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
